packages/floe-dbt/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "floe-dbt"
version = "0.1.0"
description = "dbt integration for floe-runtime"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [{ name = "Floe Team" }]
keywords = ["data", "pipeline", "dbt", "sql"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "dbt-core>=1.9",
    "floe-core",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
duckdb = ["dbt-duckdb>=1.9"]
snowflake = ["dbt-snowflake>=1.9"]
bigquery = ["dbt-bigquery>=1.9"]
postgres = ["dbt-postgres>=1.9"]
redshift = ["dbt-redshift>=1.9"]
databricks = ["dbt-databricks>=1.9"]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "hypothesis>=6.0",
]

[project.urls]
Homepage = "https://github.com/your-org/floe-runtime"
Documentation = "https://github.com/your-org/floe-runtime/tree/main/docs"
Repository = "https://github.com/your-org/floe-runtime"

[tool.hatch.build.targets.wheel]
packages = ["src/floe_dbt"]

[tool.uv.sources]
floe-core = { workspace = true }
